what i learned is how to import values to tuple "t=tuple(int(x.strip()) for x in input().split(' '))"
Given an integer, n, and n space-separated integers as input, create a tuple,t , of thosen  integers. Then compute and print the result of hash(t).

Note: hash() is one of the functions in the __builtins__ module, so it need not be imported.

Input Format

The first line contains an integer, n, denoting the number of elements in the tuple. 
The second line contains n space-separated integers describing the elements in tuple t.

Output Format

Print the result of hash(t)

Sample Input

2
1 2
Sample Output

3713081631934410656